requirements.txt
about-time==4.2.1
alive-progress==3.1.5
apturl==0.5.2
bcrypt==3.2.0
bleach==6.1.0
blinker==1.4
bokeh==3.4.1
boto3==1.34.84
botocore==1.34.84
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
copulas==0.9.2
cryptography==3.4.8
ctgan==0.9.1
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
deepecho==0.5.0
defer==1.0.6
dell-recovery==0.0.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
Faker==24.9.0
fasteners==0.14.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
future==0.18.2
gpg==1.16.0
grapheme==0.6.0
graphviz==0.20.3
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isodate==0.6.1
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcomps==0.1.15
linkify-it-py==2.0.3
lockfile==0.12.2
lorem-text==2.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.8.4
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==24.0
pandas==2.2.2
panel==1.4.1
param==2.1.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.20.0
progressbar==2.5
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
pyviz_comms==3.0.2
pyxdg==0.27
PyYAML==5.4.1
rdflib==7.0.0
rdt==1.11.0
reportlab==3.6.8
requests==2.25.1
rpm==4.17.0
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
sdmetrics==0.13.0
sdv==1.11.0
SecretStorage==3.3.1
six==1.16.0
sos==4.5.6
sympy==1.12
synthea-rdf==0.2
systemd-python==234
tenacity==8.2.3
threadpoolctl==3.4.0
torch==2.2.2
tornado==6.4
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.3
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
xyzservices==2024.4.0
zipp==1.0.0
